refactoring set amount
