+++ CommonWallet/Classes/Modules/Amount/AmountCoordinator.swift
@@ -20,8 +20,4 @@ final class AmountCoordinator: AmountCoordinatorProtocol {
         
         resolver.navigation?.push(confirmationView.controller)
     }
-
-    func close() {
-        resolver.navigation?.dismiss()
-    }
 }
