+++ CommonWallet/Classes/Modules/WithdrawAmount/WithdrawAmountProtocols.swift
@@ -5,23 +5,6 @@
 
 import IrohaCommunication
 
-protocol WithdrawAmountViewProtocol: ControllerBackedProtocol, LoadableViewProtocol, AlertPresentable {
-    func set(title: String)
-    func set(assetViewModel: AssetSelectionViewModelProtocol)
-    func set(amountViewModel: AmountInputViewModelProtocol)
-    func set(descriptionViewModel: DescriptionInputViewModelProtocol)
-    func set(feeViewModel: FeeViewModelProtocol)
-    func didChange(accessoryViewModel: AccessoryViewModelProtocol)
-}
-
-
-protocol WithdrawAmountPresenterProtocol: class {
-    func setup()
-    func confirm()
-    func presentAssetSelection()
-}
-
-
 protocol WithdrawAmountCoordinatorProtocol: CoordinatorProtocol, PickerPresentable {
     func confirm(with info: WithdrawInfo, asset: WalletAsset, option: WalletWithdrawOption)
 }
@@ -30,5 +13,5 @@ protocol WithdrawAmountCoordinatorProtocol: CoordinatorProtocol, PickerPresentab
 protocol WithdrawAmountAssemblyProtocol: class {
     static func assembleView(with resolver: ResolverProtocol,
                              asset: WalletAsset,
-                             option: WalletWithdrawOption) -> WithdrawAmountViewProtocol?
+                             option: WalletWithdrawOption) -> AmountViewProtocol?
 }
